+++ core/User/routes/api.php
@@ -30,9 +30,6 @@ use Core\User\Http\Controllers\Api\Admin\GroupController;
 use Core\User\Http\Controllers\Api\Admin\RoleController;
 use Core\User\Http\Controllers\Api\Admin\ScopeController;
 use Core\User\Http\Controllers\Api\Admin\ServiceController;
-use Core\User\Http\Controllers\Api\Admin\ServiceScopeController;
-use Core\User\Http\Controllers\Api\Admin\UserController;
-use Core\User\Http\Controllers\Api\Admin\UserScopeController;
 use Core\User\Http\Controllers\Api\User\NotificationController;
 use Illuminate\Support\Facades\Route;
 
@@ -42,23 +39,19 @@ Route::group([
     'middleware' => ['throttle:core:user:api_admin']
 ], function () {
 
-    Route::resource('groups', GroupController::class)->except('edit', 'create', 'show');
-    Route::resource('roles', RoleController::class)->except('edit', 'create', 'show');
-
-    Route::resource('services', ServiceController::class)->except('edit', 'create');
-    Route::get('services/{service}/scopes', [ServiceScopeController::class, 'index'])->name('services.scopes.index');
-    Route::post('services/{service}/scopes', [ServiceScopeController::class, 'assign'])->name('services.scopes.assign');
-    Route::delete('services/{service}/scopes/{scope}', [ServiceScopeController::class, 'revoke'])->name('services.scopes.revoke');
-
+    Route::resource('roles', RoleController::class)->only('index');
+    Route::resource('groups', GroupController::class)->only('index');
+    Route::resource('services', ServiceController::class)->only('index');
     Route::resource('scopes', ScopeController::class)->only('index');
 
-    Route::get('/users/{user}/scopes', [UserScopeController::class, 'index'])->name('users.scopes.index');
-    Route::post('/users/{user}/scopes', [UserScopeController::class, 'assign'])->name('users.scopes.assign');
-    Route::delete('/users/{user}/scopes/{scope}', [UserScopeController::class, 'revoke'])->name('users.scopes.revoke');
 
-    Route::delete('users/{user}/disable', [UserController::class, 'disable'])->name('users.disable');
-    Route::put('users/{id}/enable', [UserController::class, 'enable'])->name('users.enable');
-    Route::resource('users', UserController::class)->except('edit', 'create', 'destroy');
+    //Route::get('/users/{user}/scopes', [UserScopeController::class, 'index'])->name('users.scopes.index');
+    //Route::post('/users/{user}/scopes', [UserScopeController::class, 'assign'])->name('users.scopes.assign');
+    //Route::delete('/users/{user}/scopes/{scope}', [UserScopeController::class, 'revoke'])->name('users.scopes.revoke');
+
+    //Route::delete('users/{user}/disable', [UserController::class, 'disable'])->name('users.disable');
+    //Route::put('users/{id}/enable', [UserController::class, 'enable'])->name('users.enable');
+    //Route::resource('users', UserController::class)->except('edit', 'create', 'destroy');
 });
 
 Route::group([