+++ core/User/routes/web.php
@@ -36,9 +36,9 @@ Route::group([
 
     Route::get("/", [HomePageController::class, 'dashboard'])->name('dashboard');
 
-    Route::middleware(['password.confirm'])->group(function () {
+    Route::get("/notifications", [UserController::class, 'notifications'])->name('notifications');
 
-        Route::get("/notifications", [UserController::class, 'notifications'])->name('notifications');
+    Route::middleware(['password.confirm'])->group(function () {
         Route::get("/update", [UserController::class, 'profile'])->name('profile');
         Route::put("/update", [UserController::class, 'personalInformation'])->name('update');
         Route::get("/change-password", [UserController::class, 'formToChangePassword'])->name("password");